requirements.txt
asgiref==3.3.1
Django==3.1.5
django-constance==2.8.0
django-picklefield==3.0.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
numpy==1.19.5
pandas==1.2.1
psycopg2-binary==2.8.6
PyJWT==2.0.1
python-dateutil==2.8.1
pytz==2020.5
six==1.15.0
SQLAlchemy==1.3.22
sqlparse==0.4.1
